Fix neighbor_list_from_ase(compute_distances=True) crashing on periodic structures
_compute_distances multiplied the integer unit-shift matrix returned by
the neighbor-list kernel with the float cell; integer x float matmul is
not defined in torch, so the DEFAULT arguments of neighbor_list_from_ase
raised 'expected mat1 and mat2 to have the same dtype' on any structure
with periodic boundary conditions. Non-periodic structures dodge the path
(unit_shifts is None), which is presumably how it survived smoke tests.

Cast the shifts to the position dtype at the point of use. The new test
runs the default path on a periodic rocksalt supercell and checks the
returned distances against a float64 recomputation (1e-4) and the cutoff,
not merely the absence of a crash.

While auditing this path we also cross-checked the neighbor list itself:
on a dense triclinic 1,248-atom cell at 5.0 A the returned edge set is
bit-identical (shift vectors included) to an exact O(N^2) minimum-image
reference -- 25,242 edges, zero missing, zero spurious.

+++ src/matgl/ext/_alchmtk.py
@@ -217,7 +217,11 @@ def _compute_distances(
     """Compute the distances between the source and destination atoms."""
     vectors = positions[dst_id] - positions[src_id]
     if cell is not None and unit_shifts is not None:
-        vectors += unit_shifts @ cell
+        # unit_shifts arrives as an integer tensor from the neighbor-list
+        # kernel; integer x float matmul raises "expected mat1 and mat2 to
+        # have the same dtype", so every periodic call with
+        # compute_distances=True (the default) crashed here.
+        vectors += unit_shifts.to(positions.dtype) @ cell
     return torch.linalg.norm(vectors, dim=1)
